// lib/_TC/_SeD/_V/enums/BrdBlm_TC_SeD_V_eSlidingSystem.ts
/** ---------------------------------------------------------------------------
 * @module [BrdBlm_TC_SeD_V]
 * @author [APG] Angeli Paolo Giusto
 * @version 0.1 APG 20231108
 * @version 0.2 APG 20231227 Modulo BrdBlm server side
 * ----------------------------------------------------------------------------
 */

/**
 * Tipo guide scorrimento portone sezionale come da come da colonna ID della tabella
 * [Tabelle configurazioni porte] -> [Scorrimenti]
 */

export enum BrdBlm_TC_SeD_V_eSlidingSystem {
  S1 = "S1",
  S2 = "S2",
  S2I = "S2I",
  S2R = "S2R",
  S2RI = "S2RI",
  S3 = "S3",
  S3I = "S3I",
  S3R = "S3R",
  S4 = "S4",
  S4R = "S4R",
  SB2 = "SB2",
  SD = "SD",
  SDB = "SDB",
  SDI = "SDI",
  SL2 = "SL2",
  SL2I = "SL2I",
  SL4 = "SL4",
  SLO = "SLO",
  SLOI = "SLOI",
  SMB1 = "SMB1",
  SMB2 = "SMB2",
  SMB3 = "SMB3",
  SO = "SO",
  SR = "SR",
  STO = "STO"
}

/*! ---------------------------------------------------------------------------
 * @copyright Breda Sistemi industriali S.p.A., 2023 - http://bredasys.com
 * All rights reserved
 * @licence You cannot host, display, distribute or share this Work in any
 * form, both physical and digital. You cannot use this Work in any commercial
 * or non-commercial product, website or project. You cannot sell this Work
 * and you cannot mint an NFTs out of it.
 * ---------------------------------------------------------------------------
 */